refactor: 팀 인원 수 type 변경 (Long -> int)

+++ src/main/java/com/group/attendancemanagementsystem/domain/team/Team.java
@@ -22,7 +22,7 @@ public class Team {
     @Column(nullable = false)
     private String name;
     private String manager;
-    private Long memberCount;
+    private int memberCount;
 
     @OneToMany(mappedBy = "team")
     private List<Employee> employees = new ArrayList<>();
